+++ checkers/testdata/timeComparisonSimplify/negative_tests.go
@@ -1,4 +1,4 @@
-package timeRangeExprSimplify
+package timeComparisonSimplify
 
 import (
 	"time"
